feat(app): adding snow plugin
