+++ package.json
@@ -11,6 +11,8 @@
   "author": "",
   "license": "Apache-2.0",
   "dependencies": {
+    "axios": "^0.19.0",
+    "crypto-js": "^3.1.9-1",
     "express": "^4.17.0",
     "express-ws": "^4.0.0",
     "minimist": "^1.2.0",
